# config/swagger.yaml
swagger: '2.0'
info:
  title: TDA-TAX-2018
  version: 0.0.1
host: localhost
schemes:
  - http
basePath: /api/v1
produces:
  - application/json
paths:
  /accounts:
    get:
      summary: Get accounts
      tags:
        - Accounts
      parameters:
        - name: limit
          in: query
          type: number
          format: integer
        - name: offset
          in: query
          type: number
          format: integer
      responses:
        '200':
          description: Success
          schema:
            type: object
            properties:
              persons:
                type: array
                items:
                  $ref: '#/definitions/Account'
        '500':
          description: Server error
    post:
      summary: Create account
      tags:
        - Accounts
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/Account'
      responses:
        '201':
          description: Account created
          schema:
            $ref: '#/definitions/Account'
        '409':
          description: Already exist
        '500':
          description: Server error
  '/accounts/{id}':
    get:
      summary: Get account by id
      tags:
        - Accounts
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        '200':
          description: account item
          schema:
            $ref: '#/definitions/Account'
        '404':
          description: Not found
        '500':
          description: Server error
    put:
      summary: Update account
      tags:
        - Accounts
      parameters:
        - name: id
          in: path
          required: true
          type: string
        - name: body
          in: body
          schema:
            $ref: '#/definitions/Account'
      responses:
        '200':
          description: Account updated
          schema:
            $ref: '#/definitions/Account'
        '500':
          description: Server error 
...
definitions:
 Account:
    type: object
    properties:
      id:
        type: number
        format: integer
      login:
        type: string
      password:
        type: string
      created:
        type: string
        format: date-time
      role:
        type: string
      isBlocked:
        type: boolean
...